N501Y変異株スクリーニングの実施状況: Rename components/cards/VariantCard.vue to components/index/CardsReference/Variant/Card.vue
