+++ pages/cards/_card.vue
@@ -39,7 +39,7 @@ import DeathsByDeathDateCard from '@/components/index/CardsReference/DeathsByDea
 // 検査実施件数
 import TestedNumberCard from '@/components/index/CardsReference/TestedNumber/Card.vue'
 // N501Y変異株スクリーニングの実施状況
-import VariantCard from '@/components/cards/VariantCard.vue'
+import VariantCard from '@/components/index/CardsReference/Variant/Card.vue'
 // 新型コロナコールセンター相談件数
 import TelephoneAdvisoryReportsNumberCard from '@/components/cards/TelephoneAdvisoryReportsNumberCard.vue'
 // 受診相談窓口における相談件数
